CLI: move context git into subpackage
