+++ cmd/prescribe/cmds/context/git/legacy.go
@@ -1,4 +1,4 @@
-package context
+package git
 
 import (
 	"fmt"
@@ -9,36 +9,6 @@ import (
 	"github.com/go-go-golems/prescribe/internal/domain"
 	"github.com/spf13/cobra"
 )
-
-func NewGitCmd() (*cobra.Command, error) {
-	gitCmd := &cobra.Command{
-		Use:   "git",
-		Short: "Manage git-derived context",
-		Long:  "Manage git-derived context (history and explicit git artifacts) for the current session.",
-	}
-
-	historyCmd := &cobra.Command{
-		Use:   "history",
-		Short: "Manage derived git history",
-		Long:  "Show and configure derived git history inclusion for generation (stored in session.yaml).",
-	}
-
-	historyCmd.AddCommand(
-		newGitHistoryShowCmd(),
-		newGitHistoryEnableCmd(true),
-		newGitHistoryEnableCmd(false),
-		newGitHistorySetCmd(),
-	)
-
-	gitCmd.AddCommand(historyCmd)
-	gitCmd.AddCommand(
-		newGitContextListCmd(),
-		newGitContextRemoveCmd(),
-		newGitContextClearCmd(),
-		newGitContextAddCmd(),
-	)
-	return gitCmd, nil
-}
 
 func effectiveGitHistoryConfig(data *domain.PRData) (domain.GitHistoryConfig, bool) {
 	if data != nil && data.GitHistory != nil {
